requirements.txt
grpc==0.3.post19
numpy==1.14.0
Pillow==5.0.0
tensorflow==1.5.0
tensorflow-serving-api==1.4.0
